Moved namespace logic that is depedent of the user-service to the namespace-controller file
